fix: update misspell
