+++ .changelog/1785.txt
@@ -1,3 +1,3 @@
 ```release-note:new-resource
-tencentcloud_redis_switck_master
+tencentcloud_redis_switch_master
 ```
