+++ website/tencentcloud.erb
@@ -3200,7 +3200,7 @@
                                     <a href="/docs/providers/tencentcloud/r/redis_startup_instance_operation.html">tencentcloud_redis_startup_instance_operation</a>
                                 </li>
                                 <li>
-                                    <a href="/docs/providers/tencentcloud/r/redis_switck_master.html">tencentcloud_redis_switck_master</a>
+                                    <a href="/docs/providers/tencentcloud/r/redis_switch_master.html">tencentcloud_redis_switch_master</a>
                                 </li>
                                 <li>
                                     <a href="/docs/providers/tencentcloud/r/redis_upgrade_cache_version_operation.html">tencentcloud_redis_upgrade_cache_version_operation</a>
